add IOptions support for SDK provider
